+++ Aplicacion.sln
@@ -1,9 +1,13 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
-VisualStudioVersion = 17.12.35707.178 d17.12
+VisualStudioVersion = 17.12.35707.178
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Aplicacion", "Aplicacion.csproj", "{E061BE34-142B-4954-B84A-5721C8C3A05B}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Escritorio", "..\Escritorio\Escritorio.csproj", "{94320F54-A982-49C9-A502-C9D66FC6026F}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "API_REST", "..\API_REST\API_REST.csproj", "{070E0B2B-015C-4354-BB8A-7CB222C70255}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,8 +19,19 @@ Global
 		{E061BE34-142B-4954-B84A-5721C8C3A05B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E061BE34-142B-4954-B84A-5721C8C3A05B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E061BE34-142B-4954-B84A-5721C8C3A05B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{94320F54-A982-49C9-A502-C9D66FC6026F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{94320F54-A982-49C9-A502-C9D66FC6026F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{94320F54-A982-49C9-A502-C9D66FC6026F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{94320F54-A982-49C9-A502-C9D66FC6026F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{070E0B2B-015C-4354-BB8A-7CB222C70255}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{070E0B2B-015C-4354-BB8A-7CB222C70255}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{070E0B2B-015C-4354-BB8A-7CB222C70255}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{070E0B2B-015C-4354-BB8A-7CB222C70255}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {BFC8D3F0-0E90-4B9C-A525-6D57FB2115B2}
 	EndGlobalSection
 EndGlobal
