--- Application/ViewModels/LoginViewModel.cs
@@ -1,18 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace Aplicacion.Application.ViewModels
-{
-    public class LoginViewModel
-    {
-
-        [Required(ErrorMessage = "El correo electrónico es obligatorio.")]
-        [EmailAddress(ErrorMessage = "Debe ingresar un correo válido")]
-        public string Email { get; set; }
-
-        [Required(ErrorMessage = "La contraseña es obligatoria")]
-        [DataType(DataType.Password)]
-        public string Password { get; set; }
-
-        public bool RememberMe { get; set; }
-    }
-}
