+++ Controllers/AdminController.cs
@@ -1,0 +1,18 @@
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
+
+namespace Aplicacion.Controllers
+{
+    [Authorize(Roles = "Administrador")]
+    public class AdminController : Controller
+    {
+        public IActionResult Index()
+        {
+            return View();
+        }
+        public IActionResult Config()
+        {
+            return View();
+        }
+    }
+}
