+++ Controllers/HomeController.cs
@@ -1,0 +1,44 @@
+using System.Diagnostics;
+using Aplicacion.Core.Models;
+using Aplicacion.Models.Interfaces;
+using Microsoft.AspNetCore.Mvc;
+
+namespace Aplicacion.Controllers
+{
+    public class HomeController : Controller
+    {
+        private readonly ILogger<HomeController> _logger;
+        private readonly IProductoService _productoService;
+        public HomeController(ILogger<HomeController> logger, IProductoService productoService)
+        {
+            _logger = logger;
+            _productoService = productoService;
+        }
+
+        public async Task<IActionResult> Index()
+        {
+            var limitedProducts = await _productoService.GetLimitedProducts();
+            return View(limitedProducts);
+        }
+
+        public IActionResult Privacy()
+        {
+            return View();
+        }
+
+        public IActionResult Contact()
+        {
+            return View();
+        }
+        public IActionResult AboutUs()
+        {
+            return View();
+        }
+
+        [ResponseCache(Duration = 0, Location = ResponseCacheLocation.None, NoStore = true)]
+        public IActionResult Error()
+        {
+            return View(new ErrorViewModel { RequestId = Activity.Current?.Id ?? HttpContext.TraceIdentifier });
+        }
+    }
+}
