+++ Controllers/PedidoController.cs
@@ -1,0 +1,34 @@
+﻿using Aplicacion.Models.ViewModels;
+using Microsoft.AspNetCore.Mvc;
+using System.Text.Json;
+
+namespace Aplicacion.Controllers
+{
+    public class PedidoController : Controller
+    {
+        public IActionResult Confirmar()
+        {
+            return View();
+        }
+
+        [HttpPost]
+        public IActionResult Confirmar(string cartItemsInput)
+        {
+            System.Diagnostics.Debug.WriteLine("Datos recibidos en Confirmar: " + cartItemsInput);
+
+            if (string.IsNullOrEmpty(cartItemsInput))
+            {
+                TempData["Error"] = "No se recibieron datos del carrito.";
+                return RedirectToAction("Index", "Catalogo");
+            }
+
+            var options = new JsonSerializerOptions { PropertyNameCaseInsensitive = true };
+            var pedido = JsonSerializer.Deserialize<List<ProductoViewModel>>(cartItemsInput, options);
+
+            return View(pedido);
+        }
+
+
+
+    }
+}
