--- Core/Models/ErrorViewModel.cs
@@ -1,9 +1,0 @@
-namespace Aplicacion.Core.Models
-{
-    public class ErrorViewModel
-    {
-        public string? RequestId { get; set; }
-
-        public bool ShowRequestId => !string.IsNullOrEmpty(RequestId);
-    }
-}
