+++ Data/ApplicationDbContext.cs
@@ -1,0 +1,16 @@
+﻿using Aplicacion.Models;
+using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore;
+
+namespace Aplicacion.Data
+{
+    public class ApplicationDbContext : IdentityDbContext
+    {
+        public ApplicationDbContext(DbContextOptions<ApplicationDbContext> options) : base(options) { }
+        public DbSet<Producto> Productos { get; set; }
+        public DbSet<Categoria> Categorias { get; set; }
+        public DbSet<Servicio> Servicios { get; set; }
+        public DbSet<TipoServicio> TiposServicios { get; set; }
+
+    }
+}
