--- Infrastructure/Data/ApplicationDbContext.cs
@@ -1,29 +1,0 @@
-﻿using Aplicacion.Core.Models;
-using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore;
-
-namespace Aplicacion.Infrastructure.Data
-{
-    public class ApplicationDbContext : IdentityDbContext
-    {
-        public ApplicationDbContext(DbContextOptions<ApplicationDbContext> options) : base(options) { }
-        public DbSet<Producto> Productos { get; set; }
-        public DbSet<Categoria> Categorias { get; set; }
-
-        protected override void OnModelCreating(ModelBuilder modelBuilder)
-        {
-            base.OnModelCreating(modelBuilder); // Asegura la configuración de Identity
-
-            modelBuilder.Entity<Producto>().ToTable("Productos");
-            modelBuilder.Entity<Categoria>().ToTable("Categorias");
-
-            modelBuilder.Entity<Producto>()
-                .Property(p => p.Precio)
-                .HasPrecision(18, 2);
-
-            modelBuilder.Entity<Producto>()
-                .Property(p => p.CostoUnitario)
-                .HasPrecision(18, 2);
-        }
-    }
-}
