+++ Models/ErrorViewModel.cs
@@ -1,0 +1,9 @@
+namespace Aplicacion.Models
+{
+    public class ErrorViewModel
+    {
+        public string RequestId { get; set; }
+
+        public bool ShowRequestId => !string.IsNullOrEmpty(RequestId);
+    }
+}
