+++ Models/ViewModels/ServicioViewModel.cs
@@ -1,0 +1,6 @@
+﻿namespace Aplicacion.Models.ViewModels
+{
+    public class ServicioViewModel
+    {
+    }
+}
