--- Presentation/Controllers/AdminController.cs
@@ -1,18 +1,0 @@
-﻿using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Mvc;
-
-namespace Aplicacion.Presentation.Controllers
-{
-    [Authorize(Roles = "Administrador")]
-    public class AdminController : Controller
-    {
-        public IActionResult Index()
-        {
-            return View();
-        }
-        public IActionResult Config()
-        {
-            return View();
-        }
-    }
-}
