+++ Views/Account/Register.cshtml
@@ -1,4 +1,5 @@
-﻿@model Aplicacion.Application.ViewModels.LoginViewModel
+﻿@using Aplicacion.ViewModels
+@model LoginViewModel
 @{
     ViewData["Title"] = "Registro";
     Layout = null;
