+++ Views/Pedido/Confirmar.cshtml
@@ -1,4 +1,5 @@
-﻿@model List<Aplicacion.Application.ViewModels.ProductoViewModel>
+﻿@using Aplicacion.ViewModels
+@model List<ProductoViewModel>
 @{
     ViewData["Title"] = "Resumen del Pedido";
 }
