+++ Views/Producto/Index.cshtml
@@ -1,4 +1,5 @@
 ﻿@using Aplicacion.Application.ViewModels
+@using Aplicacion.ViewModels
 @model IEnumerable<ProductoViewModel>
 
 @{
