+++ Views/Shared/Error.cshtml
@@ -1,4 +1,5 @@
 ﻿@using Aplicacion.Core.Models
+@using Aplicacion.Models
 @model ErrorViewModel
 @{
     ViewData["Title"] = "Error";
